+++ terraform/core/main.tf
@@ -18,6 +18,11 @@ resource "random_password" "s3_password" {
   special          = false
 }
 
+resource "random_password" "service_password" {
+  length           = 32
+  special          = false
+}
+
 resource "random_password" "oidc_session_secret" {
   length           = 32
   special          = false